requirements.txt
argcomplete==3.6.2
bleak==0.22.3
certifi==2025.8.3
charset-normalizer==3.4.3
dbus-fast==2.44.5
dotmap==1.3.30
idna==3.10
meshtastic==2.7.3
nicegui==1.4.25
packaging==24.2
print-color==0.4.6
protobuf==6.32.1
Pypubsub==4.0.3
PyQRCode==1.2.1
pyserial==3.5
pytap2==2.3.0
PyYAML==6.0.3
requests==2.32.5
tabulate==0.9.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.14
